:bento: :recycle: reformat code hide unuse class outside package
