+++ lib/src/time_func.dart
